Fixes refactor

+++ DeliveryApi/Services/BasketService/BasketService.cs
@@ -16,8 +16,7 @@ public class BasketService : IBasketService
     }
     public async Task<List<BasketDTO>> GetUserBasket(string token)
     {
-        var email = JwtParseHelper.GetClaimValue(token, ClaimTypes.Email);
-        var user = await _context.User.FirstOrDefaultAsync(u => u.Email == email);
+        var user = await JwtParseHelper.GetUserFromContext(token, _context);
         var userBasket = _context.Basket.Where(b => b.UserId == user.Id).Include(b => b.Dish).ToList();
         List<BasketDTO> userBasketDTO = new List<BasketDTO>();
         foreach (var item in userBasket)
@@ -38,10 +37,9 @@ public class BasketService : IBasketService
 
     public async Task AddToUserBasket(string token, Guid dishId)
     {
-        var email = JwtParseHelper.GetClaimValue(token, ClaimTypes.Email);
-        var user = await _context.User.FirstOrDefaultAsync(u => u.Email == email);
+        var user = await JwtParseHelper.GetUserFromContext(token, _context);
         var dishCard = await _context.Basket.FirstOrDefaultAsync(b => b.UserId == user.Id && b.DishId == dishId);
-
+        
         if (dishCard == null)
         {
             await _context.Basket.AddAsync(new Basket { UserId = user.Id, DishId = dishId, Amount = 1 });
@@ -56,10 +54,12 @@ public class BasketService : IBasketService
 
     public async Task DeleteFromUserBasket(string token, Guid dishId, bool increase)
     {
-        var email = JwtParseHelper.GetClaimValue(token, ClaimTypes.Email);
-        var user = await _context.User.FirstOrDefaultAsync(u => u.Email == email);
+        var user = await JwtParseHelper.GetUserFromContext(token, _context);
         var dishCard = await _context.Basket.FirstOrDefaultAsync(b => b.UserId == user.Id && b.DishId == dishId);
-
+        if (dishCard == null)
+        {
+            throw new Exception(message: $@"such dish with id {dishId} not found");
+        }
         if (increase == false || dishCard.Amount == 1)
         {
             _context.Basket.Remove(dishCard);
